requirements.txt
PyQt6>=6.7,<7
Pillow>=11.0,<12
piexif>=1.1,<2
Send2Trash>=1.8,<2
ruff>=0.11,<0.12
